pyproject.toml
[project]
name = "grpc-kv-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cffi>=1.17.1",
    "cryptography>=45.0.3",
    "google>=3.0.0",
    "grpcio>=1.71.0",
    "protobuf>=6.31.1",
    "rich>=14.0.0",
]
